refactor: Remove unused getAnsweredQuestionsCount() function
The getAnsweredQuestionsCount() function was previously used to calculate the number of answered questions in the quiz, but it is no longer necessary for our current use case. We only need to display the overall percentage score to the user at the end of the quiz. Therefore, this commit removes the unused function, simplifying the code and improving code readability.

As a result, the code is now cleaner and more focused on displaying the correct score percentage to the user.